requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
reportlab
PyMuPDF
